refactor(onboard): extract policy selection prompts
